Deployment/Lambda/requirements.txt
requests==2.31.0
boto3==1.34.0
botocore==1.34.0
python-dateutil==2.8.2 